+++ tests/roots/test-ext-autodoc-typevar/conf.py
@@ -1,0 +1,3 @@
+extensions = ['sphinx.ext.autodoc']
+
+nitpicky = True